Add files via upload
The data indicates that black people were affected worse than whites.  There are many factors one may not consider are significant in determining the access to healthcare one may receive or the outcome of their health problems.  Different groups of people do not receive equal care, and the pandemic highlighted several areas where this is true.  During the pandemic, there was a large disparity in the death rates of the greater population compared to African Americans.  They had a much higher death rate.  People in prison did not receive the proper amenities to deal with COVID-19.  Economic status is a major factor in the outcome of health issues.